// light-bootstrap-dashboard-angular2-master/src/app/Model/Splunk.ts
import { CommandePasserPar } from "./CommandePasserPar";


export class Splunk {
    splunkid: number;
    client: string;
     licences: {
      nomDesLicences: string;
      quantite: string;
      dateEx: Date;
}[];
    dureeLicence: string;
    nomDuContact: string;
    adresseEmailContact: string;
    mailAdmin: string;
    commandePasserPar: CommandePasserPar;
    ccMail: string[];
    numero: string;
    sousContrat: boolean;
    approuve?: boolean;
    remarques: string;
    fichier?: string;
    fichierOriginalName?: string;
}